+++ dependancy_installer.py
@@ -1,0 +1,4 @@
+import os
+
+
+os.system("pip install -r requirements.txt")
