+++ requirements.txt
@@ -1,2 +1,3 @@
 requests>=2.31.0
 PySide6>=6.6.2
+qt_material
